pyk/: move integration tests into unit tests
